(IN-97)feat: add sheduler for deleting avatar

+++ apps/app/src/sheduler/events-sheduler.ts
@@ -6,8 +6,13 @@ import { EventsService } from '../features/rmq-provider/events-db/events.service
 import { $Enums } from '../../prisma/client';
 
 import { Data, HandleEventForProfileAvatarCommand } from './command/handle-event-for-profile-avatar.command';
+import {
+  DataForDelete,
+  HandleEventForDeleteProfileAvatarCommand,
+} from './command/handle-event-for-delete-profile-avatar.command';
 
 import Entity = $Enums.Entity;
+import EventType = $Enums.EventType;
 
 @Injectable()
 export class EventsSheduler {
@@ -21,9 +26,19 @@ export class EventsSheduler {
     const events = await this.eventsService.getResolvedEvents();
 
     events.forEach((e) => {
-      if (e.entity === Entity.PROFILE) {
+      // если event касается загрузки аватарки
+      if (e.entity === Entity.PROFILE && e.eventType === EventType.CREATE) {
         this.commandBus.execute(new HandleEventForProfileAvatarCommand(e as Data));
         this.eventsService.deleteEvent(e.id);
+      }
+      // если event касается удаления аватарки
+      if (e.entity === Entity.PROFILE && e.eventType === EventType.DELETE) {
+        try {
+          this.commandBus.execute(new HandleEventForDeleteProfileAvatarCommand(e as DataForDelete));
+        } catch {
+          // TODO logger
+          console.error('error in event sheduler');
+        }
       }
     });
   }
